Added dashboard design layout
